refactor: internal/sql — split fold.go → migrations.go (v0.8.1 Part A)
Move migration-directory detection (migrationPatterns,
classifyMigrationName, minMigrationFiles, IsMigrationDir) out of
fold.go into a sibling migrations.go in the same package. fold.go
keeps the replay machinery — splitStatements / parseAlterTable /
applyAlter / foldState — that drives the merged-state chunks.

Zero behavior change. Same package boundary; no caller changes.
The "regexp" import drops out of fold.go since the patterns live
in migrations.go now.

Load-bearing for v0.8.1 Part C: RENAME COLUMN / RENAME CONSTRAINT
folding (closing #14) will land its per-table rename map + chain
resolution helpers in migrations.go rather than further bloating
fold.go.
